player_back/player_back/users/serializers.py
# Changes
from django.contrib.auth.models import Group
from player_back.users.models import User


from rest_framework import serializers
from django.contrib.auth.hashers import make_password
from rest_framework_simplejwt.tokens import AccessToken, RefreshToken, TokenError
from django.utils.text import gettext_lazy as _

class RefreshTokenSerializer(serializers.Serializer):
    refresh = serializers.CharField()
    # access = serializers.CharField()

    default_error_messages = {
        'bad_token': _('Token is invalid or expired')
    }

    def validate(self, attrs):
        self.refresh_token = attrs['refresh']
        # self.access_token = attrs['access']
        return attrs

    def save(self, **kwargs):
        try:
            # AccessToken(self.access_token).blacklist()
            RefreshToken(self.refresh_token).blacklist()
        except TokenError:
            self.fail('bad_token')

class UserSerializer(serializers.ModelSerializer):
    def validate_password(self, value: str) -> str:
        return make_password(value)

   
    class Meta:
        model = User
        fields = ['id', 'username', 'password', 'email', 'mode', 'groups']
         # Changes password to write only, user never be able to access it
        extra_kwargs = {
            'email': {'required': True},
            'password': {'write_only': True},
        }


class GroupSerializer(serializers.ModelSerializer):
    class Meta:
        model = Group
        fields = ['id', 'name']