player_back/requirements.txt
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
psycopg2-binary==2.9.3
PyJWT==2.3.0